Enhance actors API with error handling and validation; update Swagger host to use environment variable

+++ controllers/actors.js
@@ -1,34 +1,52 @@
+const createError = require('http-errors');
 const mongodb = require('../data/database');
 const ObjectId = require('mongodb').ObjectId;
 
-const getAll = async (req, res) => {
-    //#swagger.tags=['Actors']
-    const response = await mongodb.getDatabase().db().collection("actors").find();
-    response.toArray().then((movies) => {
-        res.setHeader("Content-Type", "application/json");
-        res.status(200).json(movies);
-    })
+const getAll = async (req, res, next) => {
+    try {
+        //#swagger.tags=['Actors']
+        const response = await mongodb.getDatabase().db().collection("actors").find();
+        if(!response) throw createError(404, "Not Actors found");
+
+        response.toArray().then((movies) => {
+            res.setHeader("Content-Type", "application/json");
+            res.status(200).json(movies);
+        })
+    } catch (err) {
+        next(err);
+    }
 };
 
-const getSingle = async (req, res) => {
-    //#swagger.tags=['Actors']
-    const actorsId = new ObjectId(req.params.id);
-    const response = await mongodb.getDatabase().db().collection("actors").findOne({_id: actorsId});
-    res.status(200).json(response);
-}
+const getSingle = async (req, res, next) => {
+    try {
+        //#swagger.tags=['Actors']
+        const actorId = new ObjectId(req.params.id);
+        
+        const validID = await validateExistingID(actorId);
+        if(!validID) throw createError(404, "Actors ID does not exist");
 
-const getSingleQueries = async (req, res) => {
-    //#swagger.tags=['Actors']
-    const actorsName = String(req.query.name);
-    const response = await mongodb.getDatabase().db().collection("actors").findOne({ name: actorsName });
-    if (response) {
+        const response = await mongodb.getDatabase().db().collection("actors").findOne({_id: actorId});
         res.status(200).json(response);
-    } else {
-        res.status(404).json({ message: "Movie not found." });
+    } catch (err) {
+        next(err);
     }
 }
 
-const addActors = async (req, res) => {
+const getSingleQueries = async (req, res, next) => {
+    try {
+        //#swagger.tags=['Actors']
+        const actorsName = String(req.query.name);
+        const response = await mongodb.getDatabase().db().collection("actors").findOne({ name: actorsName });
+
+        if(!response) throw createError(404, "Actors Name does not exist");
+        
+        return res.status(200).json(response);
+    } catch (err) {
+        next(err);
+    }
+}
+
+const addActors = async (req, res, next) => {
     //#swagger.tags=['Actors']
     const actor = {
         name: req.body.name,
@@ -40,41 +58,63 @@ const addActors = async (req, res) => {
     }
     const response = await mongodb.getDatabase().db().collection("actors").insertOne(actor);
     if (response.acknowledged) {
-        res.status(204).send();
+        return res.status(204).send();
     } else {
         res.status(500).json(response.error || "Some error occurred while adding the movie.")
     }
 }
 
-const updateActors = async (req, res) => {
+const updateActors = async (req, res, next) => {
     //#swagger.tags=['Actors']
-    const actorId = new ObjectId(req.params.id);
-    const actor = {
-        name: req.body.name,
-        last_name: req.body.last_name,
-        birthdate: req.body.birthdate,
-        height_m: req.body.height_m,
-        country: req.body.country,
-        movies_id: req.body.movies_id
-    }
-    const response = await mongodb.getDatabase().db().collection("actors").replaceOne({_id: actorId}, actor);
-    if (response.modifiedCount > 0) {
-        res.status(204).send();
-    } else {
-        res.status(500).json(response.error || "Some error occurred while updating the movie.")
+    try {
+        const actorId = new ObjectId(req.params.id);
+        const actor = {
+            name: req.body.name,
+            last_name: req.body.last_name,
+            birthdate: req.body.birthdate,
+            height_m: req.body.height_m,
+            country: req.body.country,
+            movies_id: req.body.movies_id
+        }
+
+        const validID = await validateExistingID(actorId);
+
+        if(!validID) throw createError(404, "Actors ID does not exist");
+    
+        const response = await mongodb.getDatabase().db().collection("actors").replaceOne({_id: actorId}, actor);
+        if (response.modifiedCount > 0) {
+            res.status(204).send();
+        }
+    } catch (err) {
+        next(err);
     }
 };
 
-const deleteActors = async (req, res) => {
-    //#swagger.tags=['Actors']
-    const actorId = new ObjectId(req.params.id);
-    const response = await mongodb.getDatabase().db().collection("actors").deleteOne({_id: actorId});
-    if (response.deletedCount > 0) {
-        res.status(204).send();
-    } else {
-        res.status(500).json(response.error || "Some error occurred while deleting the movie.")
+const deleteActors = async (req, res, next) => {
+    try {
+        //#swagger.tags=['Actors']
+        const actorId = new ObjectId(req.params.id);
+
+        const validID = await validateExistingID(actorId);
+
+        if(!validID) throw createError(404, "Actors ID does not exist");
+        
+        const response = await mongodb.getDatabase().db().collection("actors").deleteOne({_id: actorId});
+        if (response.deletedCount > 0) {
+            res.status(204).send();
+        }
+    } catch (err) {
+        next(err);
     }
+
+
 };
+
+
+const validateExistingID = async (id) => {
+    const actorsID = await mongodb.getDatabase().db().collection("actors").findOne({_id: id});
+    return !!actorsID;
+}
 
 module.exports = {
     getAll,
